Adding edit paragraoh function
